+++ definition/user-tasks/routes/paths/self/tasks/post.@.js
@@ -4,6 +4,11 @@ export const description = 'Create a task that is owned by a single user.'
 
 export const tags = [ 'userTasks' ]
 
+export const security = [
+	{ cookie: [] },
+	{ api: [] },
+]
+
 export const requestBody = {
 	description: 'Create a task.',
 	content: {